Fix reflection on top level declarations from other modules
The main problem is that moduleName that is being passed to KPackageImpl is
useless: as can be seen in
ClosureCodegen.generateCallableReferenceDeclarationContainer, the name of the
current module is always written to the class file for a callable reference,
not the name of the module of the referenced declaration. To get the name of
the module of the referenced declaration, we can parse header from the given
Class object and extract the package_module_name protobuf extension. Similar
code was already there in ReflectJvmMapping.kt but it did not support
multi-file classes, of which there are a lot in the standard library; this is
now supported

 #KT-12630 Fixed
 #KT-14731 Fixed